Move LoadingStopsErrorRow & LoadingStopsRow to generic components
