+++ ui/src/components/stop-registry/search/components/index.ts
@@ -1,1 +1,3 @@
+export * from './LoadingStopsErrorRow';
+export * from './LoadingStopsRow';
 export * from './StopGroupSelector';
